Update deid_recipe.txt
There was a crucial variable missing from this file: PatientSex. 
When sharing de-identified patient dicom files for research (or other) purposes, you must remove ALL PHI. Sex is considered PHI. If you don't add this variable to the deid_recipe, the letter M or F remains on the dicom files.